refactor: 도메인에 맞춰 코드 수정
